Qt import and usability refactoring
Force the window to appear on top when started.

Finally fixed the stuck-in-pan/CTRL permanently set issue. Turns out to
be an issue with ginga which has been reported and fixed.

Refactored how ginga is imported, including its Qt helpers module.

Fixed deleselection of drawing objects when focus is lost.

Resolves #118 (again)